Moved SafeScreen file into components folder
